backend/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Mail==0.10.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
load-dotenv==0.1.0
MarkupSafe==3.0.2
packaging==24.2
python-dotenv==1.0.1
Werkzeug==3.1.3
